FEXCore: Add support for getting signal and jit region information
Signal information needs to be moved to the frontend once the rest of
the signal handling gets moved there. So I left a comment about it.

The cooperative backtrace service needs to have both host and guest side
signaling information, so we need to pass that information over.

With JIT information we can tell if a signal happened in guest or host
code. Combining information from both the guest and host will give us a
full view of why something has crashed.